Functionalities.

+++ backend/src/main/java/com/crypto/simulator/backend/model/CryptoHolding.java
@@ -3,34 +3,34 @@ package com.crypto.simulator.backend.model;
 import java.math.BigDecimal;
 
 public class CryptoHolding {
-    private Long id;
-    private Long userId;
+    private Integer id;
+    private Integer userId;
     private String cryptoSymbol;
     private BigDecimal quantity;
 
     public CryptoHolding() {
     }
 
-    public CryptoHolding(Long id, Long userId, String cryptoSymbol, BigDecimal quantity) {
+    public CryptoHolding(Integer id, Integer userId, String cryptoSymbol, BigDecimal quantity) {
         this.id = id;
         this.userId = userId;
         this.cryptoSymbol = cryptoSymbol;
         this.quantity = quantity;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return this.id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
-    public Long getUserId() {
+    public Integer getUserId() {
         return this.userId;
     }
 
-    public void setUserId(Long userId) {
+    public void setUserId(Integer userId) {
         this.userId = userId;
     }
